Begin AI Rework
Separated AI code into 2 files for data and logic
created 8 files for AI presets according to its Cell.enmTeam
