+++ Dots-RTS-Development/Assets/Scripts/AI.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 7a5f8e9401a637e43b90e9d744e0fe61
+folderAsset: yes
+timeCreated: 1506673248
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
