+++ Dots-RTS-Development/Assets/Scripts/AI/AI_Behaviour.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 3a2ef4e139294a74baf360354015977a
+timeCreated: 1506673248
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
