+++ Dots-RTS-Development/Assets/Scripts/AI/Difficulty/AI_0.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: a1611c62423d3724e84d7c9619f3aeac
+timeCreated: 1506677100
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
