+++ Dots-RTS-Development/Assets/Scripts/AI/Difficulty/AI_1.cs
@@ -1,0 +1,6 @@
+﻿using UnityEngine;
+using System.Collections;
+
+public class AI_1 : AI_Behaviour {
+	
+}
