+++ Dots-RTS-Development/Assets/Scripts/AI/Difficulty/AI_1.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 38daf636a4ff1154882192bbf279b5eb
+timeCreated: 1506681358
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
