+++ Dots-RTS-Development/Assets/Scripts/AI/Difficulty/AI_2.cs
@@ -1,0 +1,6 @@
+﻿using UnityEngine;
+using System.Collections;
+
+public class AI_2 : AI_Behaviour {
+	
+}
