+++ Dots-RTS-Development/Assets/Scripts/AI/Difficulty/AI_2.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: aed1e7cf739fc034a971791f04cfd5a7
+timeCreated: 1506681358
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
