+++ Dots-RTS-Development/Assets/Scripts/AI/Difficulty/AI_3.cs
@@ -1,0 +1,6 @@
+﻿using UnityEngine;
+using System.Collections;
+
+public class AI_3: AI_Behaviour {
+	
+}
