+++ Dots-RTS-Development/Assets/Scripts/AI/Difficulty/AI_5.cs
@@ -1,0 +1,6 @@
+﻿using UnityEngine;
+using System.Collections;
+
+public class AI_5 : AI_Behaviour {
+	
+}
