+++ Dots-RTS-Development/Assets/Scripts/AI/Difficulty/AI_5.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 70313f63b9ae8074aa0b3e29a162cc1d
+timeCreated: 1506681358
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
