+++ Dots-RTS-Development/Assets/Scripts/AI/Difficulty/AI_6.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 590e1efbaeb9ab04ca85f46913245ce1
+timeCreated: 1506681358
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
